[디노] step5-3 스프링 카페 6단계 - 댓글 (#52)
* Refactor: 중복된 어노테이션 제거

* Feat: 변경된 유저를 세션에 다시 저장하도록 수정

* Test: 비밀번호 틀릴 시 에러 확인하도록 수정

* Refactor: Article 기본 생성자 변경

* Feat: Reply 엔티티 생성

* Refactor: getCreatedAt 수정

* Feat: ReplyRepository와 구현체 생성

* Refactor: Id Long으로 변경

* Fix: Id 자동 생성 추가

* Feat: post요청 보내는 기능 추가

* Refactor: int 사용하던 것을 Long으로 수정

* Feat: 컨트롤러와 서비스 생성
우선 기본적으로 Reply를 추가하는 메서드 생성

* Feat: 댓글 DTO 생성

* Feat: 유저의 id에서 이름으로 수정

* Fix: 필드명 수정

* Feat: Dto에 userId 추가

* Feat: 해당하는 article의 replyDtoList를 반환하는 메서드 생성

* Feat: article에서 reply를 보여주도록 함

* Feat: id로 reply 찾는 메서드 추가

* Feat: reply 못 찾을 때 커스텀 예외 생성

* Feat: post url 수정

* Feat: 댓글 지우기 기능 추가

* Feat: 예외 핸들러에 핸들할 예외 추가

* Feat: article에 소프트 딜리트 추가

* Fix: mustache에서 html 사용하도록 수정

* fix: 글 수정

* Refactor: 메서드명 변경

* Feat: Reply에 소프트 딜리트 추가

* Refactor: 중복 체크 로직 제거

* Refactor: Article과 Reply 양방향 관계로 수정
Reply의 생성 책임 서비스에서 Article로 이동

* Refactor: ReplyViewDto를 ArticleService에서 생성하도록 변경
추후 N+1 문제 query로 fetch join 하여 해결할 수 있을 듯

* Feat: h2 console 설정 추가

* Feat: docker 용 properties 추가

* Feat: Dockerfile 추가
멀티 스테이지 Dockerfile로, 자바 코드 빌드도 함께 하도록 함.

* Feat: createdAt을 서울 시간으로 고정
EC2를 사용해서 배포 시 서울 시간에 맞지 않아 서울 시간으로 고정

* Feat: 배포를 편리하게 하기위해 deploy.sh 파일 생성

* Fix: deploy.sh 실행권한 변경하고, db 마운팅 추가

* Feat: docker.ignore 추가

* Docs: README 작성

* Feat: 데이터 디렉토리 생성 추가

* Docs: 배포 URL 추가

* Docs: 도커 설명 추가

+++ src/main/java/codesquad/codestagram/repository/ArticleRepository.java
@@ -10,7 +10,7 @@ public interface ArticleRepository {
     Article save(Article article);
     List<Article> findByUser(User user);
     List<Article> findByTitle(String title);
-    Optional<Article> findById(int id);
+    Optional<Article> findById(Long id);
     List<Article> findAll();
     void delete(Article article);
 }
